+++ .eslintrc.js
@@ -8,10 +8,11 @@ module.exports = {
   overrides: [
     ...overrides,
     {
-      files: 'test/*.js',
+      files: 'test/**/*.+(t|j)s',
       rules: {
         'no-magic-numbers': 'off',
         'promise/prefer-await-to-callbacks': 'off',
+        'unicorn/filename-case': 'off',
       },
     },
   ],