+++ src/function/index.ts
@@ -1,6 +1,14 @@
-export { Context as HandlerContext } from './context'
-export { Event as HandlerEvent } from './event'
-export { Handler, HandlerCallback } from './handler'
-export { Response as HandlerResponse } from './response'
-export { getSecrets, withSecrets, getNetlifyGraphToken, GraphTokenResponse, HasHeaders } from '../lib/graph'
-export { NetlifySecrets } from '../lib/secrets_helper'
+export { Context as HandlerContext } from './context.js'
+export { Event as HandlerEvent } from './event.js'
+export { Handler, BackgroundHandler, HandlerCallback } from './handler.js'
+export { Response as HandlerResponse } from './response.js'
+export {
+  getSecrets,
+  getSecretsForBuild,
+  withSecrets,
+  getNetlifyGraphToken,
+  getNetlifyGraphTokenForBuild,
+  GraphTokenResponse,
+  HasHeaders,
+} from '../lib/graph.js'
+export { NetlifySecrets } from '../lib/secrets_helper.js'